Таск4 исправлен согласно замечаний
